Update dg field tests, move them to test directory, update dg terms test accordingly.
